Try to fix DOS malloc handler

+++ include/stddef.h
@@ -1,8 +1,8 @@
 #ifndef _STDDEF_H
 #define _STDDEF_H
 
-typedef unsigned long long size_t;
-typedef long long ptrdiff_t;
+typedef unsigned long size_t;
+typedef long ptrdiff_t;
 
 #define NULL 0
